Sort some types between sdk38 and cosmwasm
