+++ packages/cosmwasm/src/index.ts
@@ -1,6 +1,5 @@
 import * as logs from "./logs";
-import * as types from "./types";
-export { logs, types };
+export { logs };
 
 export { BroadcastMode, RestClient, TxsResponse } from "./restclient";
 export {
@@ -13,7 +12,6 @@ export {
   ContractDetails,
   CosmWasmClient,
   GetNonceResult,
-  IndexedTx,
   PostTxResult,
   SearchByHeightQuery,
   SearchByIdQuery,
@@ -22,7 +20,6 @@ export {
   SearchTxQuery,
   SearchTxFilter,
 } from "./cosmwasmclient";
-export { decodeBech32Pubkey, encodeBech32Pubkey, encodeSecp256k1Pubkey } from "./pubkey";
 export {
   ExecuteResult,
   FeeTable,
@@ -32,3 +29,11 @@ export {
   UploadMeta,
   UploadResult,
 } from "./signingcosmwasmclient";
+export {
+  isMsgExecuteContract,
+  isMsgInstantiateContract,
+  isMsgStoreCode,
+  MsgStoreCode,
+  MsgExecuteContract,
+  MsgInstantiateContract,
+} from "./msgs";
