+++ packages/sdk38/src/index.ts
@@ -4,6 +4,26 @@ export { logs, types };
 
 export { pubkeyToAddress, rawSecp256k1PubkeyToAddress } from "./address";
 export { Coin, coin, coins } from "./coins";
+
+export {
+  Account,
+  Block,
+  BlockHeader,
+  Code,
+  CodeDetails,
+  Contract,
+  ContractDetails,
+  CosmosClient,
+  GetNonceResult,
+  IndexedTx,
+  PostTxResult,
+  SearchByHeightQuery,
+  SearchByIdQuery,
+  SearchBySentFromOrToQuery,
+  SearchByTagsQuery,
+  SearchTxQuery,
+  SearchTxFilter,
+} from "./cosmosclient";
 export { unmarshalTx } from "./decoding";
 export { makeSignBytes, marshalTx } from "./encoding";
 export { BroadcastMode, RestClient, TxsResponse } from "./restclient";
@@ -11,3 +31,4 @@ export { Pen, Secp256k1Pen, makeCosmoshubPath } from "./pen";
 export { decodeBech32Pubkey, encodeBech32Pubkey, encodeSecp256k1Pubkey } from "./pubkey";
 export { findSequenceForSignedTx } from "./sequence";
 export { encodeSecp256k1Signature, decodeSignature } from "./signature";
+export { FeeTable, SigningCallback, SigningCosmosClient } from "./signingcosmosclient";
