+++ packages/sdk38/types/index.d.ts
@@ -3,6 +3,25 @@ import * as types from "./types";
 export { logs, types };
 export { pubkeyToAddress, rawSecp256k1PubkeyToAddress } from "./address";
 export { Coin, coin, coins } from "./coins";
+export {
+  Account,
+  Block,
+  BlockHeader,
+  Code,
+  CodeDetails,
+  Contract,
+  ContractDetails,
+  CosmosClient,
+  GetNonceResult,
+  IndexedTx,
+  PostTxResult,
+  SearchByHeightQuery,
+  SearchByIdQuery,
+  SearchBySentFromOrToQuery,
+  SearchByTagsQuery,
+  SearchTxQuery,
+  SearchTxFilter,
+} from "./cosmosclient";
 export { unmarshalTx } from "./decoding";
 export { makeSignBytes, marshalTx } from "./encoding";
 export { BroadcastMode, RestClient, TxsResponse } from "./restclient";
@@ -10,3 +29,4 @@ export { Pen, Secp256k1Pen, makeCosmoshubPath } from "./pen";
 export { decodeBech32Pubkey, encodeBech32Pubkey, encodeSecp256k1Pubkey } from "./pubkey";
 export { findSequenceForSignedTx } from "./sequence";
 export { encodeSecp256k1Signature, decodeSignature } from "./signature";
+export { FeeTable, SigningCallback, SigningCosmosClient } from "./signingcosmosclient";
